refactor: centralizing login an refresh method on single handler file
